# docs/swagger.yaml
basePath: /api/v1/wallet
definitions:
...
  resp.BasePageResp:
    properties:
      count:
        description: 数量
        type: integer
      currentPage:
        description: 当前页数
        type: integer
      pageSize:
        description: 每页数据条数
        type: integer
      totalPage:
        description: 总页数
        type: integer
    type: object
...
  resp.WithdrawalAddrResp:
    properties:
      address:
        description: 地址
        type: string
      chain_code:
        description: 链code
        type: string
      chain_name:
        description: 链name
        type: string
      currency_name:
        description: 货币名称
        type: string
      name:
        description: 地址名称
        type: string
      withdrawal_addr_id:
        description: 地址id
        type: integer
    type: object
...
  resp.WithdrawalDetailListResp:
    properties:
      items:
        description: 帐号详情
        items:
          $ref: '#/definitions/resp.WithdrawalDetailResp'
        type: array
      page:
        $ref: '#/definitions/resp.BasePageResp'
        description: 分页
        type: object
    type: object
  resp.WithdrawalDetailResp:
    properties:
      address:
        description: 地址
        type: string
      poundage:
        description: 手续费
        type: number
      status:
        description: 状态 0已提交,1待审核,2审核中,3通过,4不通过,5已完成,6取消
        type: integer
      symbol:
        description: 代币代号
        type: string
      type:
        description: 标识 coin,token
        type: string
      updated_at:
        description: 更新时间
        type: string
      value:
        description: 金额
        type: number
    type: object
info:
  contact: {}
  description: 测试
  license: {}
  title: 测试
  version: 0.0.1
paths:
...
  /withdrawal/addr/add:
    post:
      description: 创建提现地址
      parameters:
      - description: 链id
        in: formData
        name: block_chain_id
        required: true
        type: integer
      - description: 币种id
        in: formData
        name: currency_id
        required: true
        type: integer
      - description: 地址
        in: formData
        name: address
        required: true
        type: string
      - description: 名称
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 创建提现地址接口
      tags:
      - Withdrawal 提现功能
  /withdrawal/addr/list:
    get:
      description: 读取提币地址列表
      parameters:
      - description: 链id
        in: query
        name: block_chain_id
        required: true
        type: integer
      - description: 币种id
        in: query
        name: currency_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.WithdrawalAddrResp'
      security:
      - ApiKeyAuth: []
      summary: 读取提币地址列表接口
      tags:
      - Withdrawal 提现功能
  /withdrawal/addr/remove:
    post:
      description: 删除提现地址
      parameters:
      - description: 提现地址id
        in: query
        name: withdrawal_addr_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 删除提现地址接口
      tags:
      - Withdrawal 提现功能
  /withdrawal/addr/update:
    post:
      description: 更新提现地址
      parameters:
      - description: 提现地址id
        in: formData
        name: withdrawal_addr_id
        required: true
        type: integer
      - description: 地址
        in: formData
        name: address
        required: true
        type: string
      - description: 名称
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 更新提现地址接口
      tags:
      - Withdrawal 提现功能
  /withdrawal/callback:
    post:
      description: 提现回调
      parameters:
      - description: app_id
        in: formData
        name: app_id
        required: true
        type: string
      - description: order_id
        in: formData
        name: order_id
        required: true
        type: string
      - description: transaction_hash
        in: formData
        name: transaction_hash
        required: true
        type: string
      - description: block_number
        in: formData
        name: block_number
        required: true
        type: string
      - description: from_address
        in: formData
        name: from_address
        required: true
        type: string
      - description: to_address
        in: formData
        name: to_address
        required: true
        type: string
      - description: symbol
        in: formData
        name: symbol
        required: true
        type: string
      - description: contract_address
        in: formData
        name: contract_address
        required: true
        type: string
      - description: value
        in: formData
        name: value
        required: true
        type: string
      - description: code
        in: formData
        name: code
        required: true
        type: string
      - description: message
        in: formData
        name: message
        required: true
        type: string
      - description: hash
        in: formData
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 提现回调接口
      tags:
      - Withdrawal 提现功能
  /withdrawal/detail:
    get:
      description: 获取提现明细
      parameters:
      - description: 长度
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 页数
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.WithdrawalDetailListResp'
      security:
      - ApiKeyAuth: []
      summary: 获取提现明细接口
      tags:
      - Withdrawal 提现功能
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"